@@ -1,17 +1,13 @@
 package azeem.play.models;
 
+import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.util.Date;
 
 public class Expense {
     private String name;
     private double amount;
     private Date date;
-
-    // Constructors, getters, and setters
-
-    // Default constructor
-    public Expense() {
-    }
 
     // Parameterized constructor
     public Expense(String name, double amount, Date date) {
@@ -22,29 +18,27 @@ public class Expense {
 
     // Getters and setters for all attributes
 
-    public String getName() {
+    public String name() {
         return name;
     }
 
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public double getAmount() {
+    public double amount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
-        this.amount = amount;
-    }
-
-    public Date getDate() {
+    public Date date() {
         return date;
     }
 
-    public void setDate(Date date) {
-        this.date = date;
+    public LocalDateTime localDateTime() {
+        return date.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime();
     }
 
-    // Other methods as needed
+    public boolean similarLocalDateTimeWith(LocalDateTime dateTime) {
+        return localDateTime().getYear() == dateTime.getYear() &&
+                localDateTime().getMonth() == dateTime.getMonth() &&
+                localDateTime().getDayOfMonth() == dateTime.getDayOfMonth();
+
+    }
+
 }
