@@ -1,10 +1,11 @@
 package azeem.play.models;
 
-public abstract class WorldEntity {
+public abstract class WorldEntity implements BoundToTime {
     protected World world;
 
     public WorldEntity(World world) {
         this.world = world;
+        world.addDayPassedObserver(this);
     }
 
     public World world() {
